backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
openai==1.10.0
whisper==1.0.0
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3
python-docx==1.1.0
redis==5.0.1
celery==5.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
httpx==0.26.0
